Split shell code
