+++ cmd/genji/commands/app.go
@@ -1,0 +1,76 @@
+package commands
+
+import (
+	"os"
+
+	"github.com/genjidb/genji/cmd/genji/dbutil"
+	"github.com/genjidb/genji/cmd/genji/shell"
+	"github.com/urfave/cli/v2"
+)
+
+// NewApp creates the Genji CLI app.
+func NewApp() *cli.App {
+	app := cli.NewApp()
+	app.Name = "Genji"
+	app.Usage = "Shell for the Genji database"
+	app.EnableBashCompletion = true
+	app.Flags = []cli.Flag{
+		&cli.BoolFlag{
+			Name:  "bolt",
+			Usage: "use bolt engine",
+		},
+		&cli.BoolFlag{
+			Name:  "badger",
+			Usage: "use badger engine",
+		},
+	}
+
+	app.Commands = []*cli.Command{
+		NewInsertCommand(),
+		NewVersionCommand(),
+		NewDumpCommand(),
+		NewRestoreCommand(),
+	}
+
+	// Root command
+	app.Action = func(c *cli.Context) error {
+		useBolt := c.Bool("bolt")
+		useBadger := c.Bool("badger")
+		if useBolt && useBadger {
+			return cli.NewExitError("cannot use bolt and badger options at the same time", 2)
+		}
+
+		dbpath := c.Args().First()
+
+		if (useBolt || useBadger) && dbpath == "" {
+			return cli.NewExitError("db path required when using bolt or badger", 2)
+		}
+
+		engine := "memory"
+
+		if useBolt || dbpath != "" {
+			engine = "bolt"
+		}
+
+		if useBadger {
+			engine = "badger"
+		}
+
+		if dbutil.CanReadFromStandardInput() {
+			db, err := dbutil.OpenDB(c.Context, dbpath, engine)
+			if err != nil {
+				return err
+			}
+			defer db.Close()
+
+			return dbutil.ExecSQL(c.Context, db, os.Stdin, os.Stdout)
+		}
+
+		return shell.Run(c.Context, &shell.Options{
+			Engine: engine,
+			DBPath: dbpath,
+		})
+	}
+
+	return app
+}
